Updated migration of HR Dashboard functions in Dashboard Controllers with standard formatting
