+++ backend/ims_application/src/main/java/com/swp391/ims_application/service/imp/IDashboardService.java
@@ -17,4 +17,10 @@ public interface IDashboardService {
     List<InternProgressDTO> getInternProgress(int mentorId);
 
     List<ScheduleDTO> getUpcomingSchedules(int mentorId);
+
+    long countApplicationsInCampaign(int campaignId);
+
+    long countAvailableCampaigns();
+
+    long countAllInterns();
 }
