+++ backend/ims_application/src/main/java/com/swp391/ims_application/service/imp/IIternshipCampaignService.java
@@ -15,9 +15,5 @@ public interface IIternshipCampaignService {
 
     boolean deleteById(int id);
 
-    long countApplicationsInCampaign(int campaignId);
 
-    long countAvailableCampaigns();
-
-    long countAllInterns();
 }
